chore: 도메인 완전 변경을 위한 로직 삭제
