--- src/main/java/com/hyoguoo/paymentplatform/payment/infrastructure/repostitory/JpaPaymentOrderRepository.java
@@ -1,15 +1,0 @@
-package com.hyoguoo.paymentplatform.payment.infrastructure.repostitory;
-
-import com.hyoguoo.paymentplatform.payment.infrastructure.entity.PaymentOrderEntity;
-import java.util.Optional;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaPaymentOrderRepository extends JpaRepository<PaymentOrderEntity, Long> {
-
-    Optional<PaymentOrderEntity> findByOrderId(String orderId);
-
-    Page<PaymentOrderEntity> findAll(Pageable pageable);
-}
-
