+++ src/main/java/com/hyoguoo/paymentplatform/payment/presentation/dto/response/OrderCancelResponse.java
@@ -1,0 +1,11 @@
+package com.hyoguoo.paymentplatform.payment.presentation.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class OrderCancelResponse {
+
+    private final Long id;
+}
