+++ src/main/java/com/hyoguoo/paymentplatform/payment/presentation/dto/response/OrderConfirmResponse.java
@@ -1,0 +1,13 @@
+package com.hyoguoo.paymentplatform.payment.presentation.dto.response;
+
+import java.math.BigDecimal;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class OrderConfirmResponse {
+
+    private final String orderId;
+    private final BigDecimal amount;
+}
